refactor: rename panes package to 'layout'
