+++ src/main/java/io/github/somesourcecode/someguiapi/scene/layout/Region.java
@@ -21,7 +21,7 @@
  * OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package io.github.somesourcecode.someguiapi.scene.pane;
+package io.github.somesourcecode.someguiapi.scene.layout;
 
 import io.github.somesourcecode.someguiapi.util.Insets;
 import io.github.somesourcecode.someguiapi.scene.Parent;
